Refactor big integers in the interpreter to use RzBitVectors instead of BigInt

+++ helper/src/num.rs
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2024 Rot127 <unisono@quyllur.org>
 // SPDX-License-Identifier: LGPL-3.0-only
 
-pub fn subscript(mut n: u64) -> String {
+pub fn subscript(mut n: u32) -> String {
     if n == 0 {
         return "₀".to_string();
     }
